refactor(query-builder): move exceptions to a dedicated folder
