+++ src/AGerault/Database/QueryBuilder.php
@@ -3,6 +3,9 @@
 namespace AGerault\Framework\Database;
 
 use AGerault\Framework\Contracts\Database\QueryBuilderInterface;
+use AGerault\Framework\Database\Exceptions\NoConditionsBeforeDeleteException;
+use AGerault\Framework\Database\Exceptions\NoDataProvidedException;
+use AGerault\Framework\Database\Exceptions\UnsupportedSqlActionException;
 
 class QueryBuilder implements QueryBuilderInterface
 {
